Align Ruby with shared worker app dispatch
Ruby now follows the shared worker shape: the registry dispatches app entrypoints, and the harness lives in the root worker gem instead of a separate local gem dependency. The common runner also appends --app once because every worker language now accepts the registry flag.
