+++ cmd/dev/install.go
@@ -235,13 +235,6 @@ func installRuby(ctx context.Context) error {
 		return err
 	}
 	fmt.Println("✅ Ruby worker dependencies installed successfully!")
-
-	harnessDir := targetDir + "/harness"
-	fmt.Println("Installing Ruby harness dependencies...")
-	if err := runCommandInDir(ctx, harnessDir, "bundle", "install"); err != nil {
-		return err
-	}
-	fmt.Println("✅ Ruby harness dependencies installed successfully!")
 	return nil
 }
