+++ cmd/dev/lint_and_format.go
@@ -213,20 +213,8 @@ func lintAndFormatRubyWorker(ctx context.Context, workerDir string) error {
 		return err
 	}
 
-	harnessDir := workerDir + "/harness"
-
-	fmt.Println("Formatting Ruby harness...")
-	if err := runCommandInDir(ctx, harnessDir, "bundle", "exec", "rubocop", "-A"); err != nil {
-		return err
-	}
-
-	fmt.Println("Linting Ruby harness...")
-	if err := runCommandInDir(ctx, harnessDir, "bundle", "exec", "rubocop"); err != nil {
-		return err
-	}
-
 	fmt.Println("Type checking Ruby harness...")
-	if err := runCommandInDir(ctx, harnessDir, "bundle", "exec", "steep", "check"); err != nil {
+	if err := runCommandInDir(ctx, workerDir, "bundle", "exec", "steep", "check"); err != nil {
 		return err
 	}
