+++ workers/ruby/Rakefile
