+++ workers/ruby/apps/worker/app.rb
@@ -1,0 +1,38 @@
+# frozen_string_literal: true
+
+require 'harness'
+require 'temporalio/worker'
+require_relative '../../workerlib/kitchensink/activities'
+require_relative '../../workerlib/kitchensink/kitchen_sink'
+
+module Apps
+  module Worker
+    module_function
+
+    def app
+      Harness::App.new(
+        worker: method(:build_worker),
+        client_factory: Harness.method(:default_client_factory)
+      )
+    end
+
+    def build_worker(client, context)
+      Temporalio::Worker.new(
+        client: client,
+        task_queue: context.task_queue,
+        workflows: [KitchenSinkWorkflow],
+        activities: [
+          NoopActivity.new,
+          DelayActivity.new,
+          PayloadActivity.new,
+          RetryableErrorActivity.new,
+          TimeoutActivity.new,
+          HeartbeatActivity.new,
+          ClientActivity.new(client, err_on_unimplemented: context.err_on_unimplemented)
+        ],
+        logger: context.logger,
+        **context.worker_kwargs
+      )
+    end
+  end
+end
