+++ workers/ruby/harness.rb
