--- workers/ruby/harness/.rubocop.yml
@@ -1,37 +1,0 @@
-AllCops:
-  TargetRubyVersion: 3.3
-  NewCops: enable
-  SuggestExtensions: false
-  Exclude:
-    - "lib/harness/api/**/*"
-    - "vendor/**/*"
-    - "omes-temp-*/**/*"
-
-Style/Documentation:
-  Enabled: false
-
-Metrics/MethodLength:
-  Max: 60
-  Exclude:
-    - "tests/**/*"
-
-Metrics/ClassLength:
-  Max: 300
-
-Metrics/ModuleLength:
-  Max: 200
-
-Metrics/BlockLength:
-  Enabled: false
-
-Metrics/CyclomaticComplexity:
-  Max: 15
-
-Metrics/PerceivedComplexity:
-  Max: 15
-
-Metrics/AbcSize:
-  Max: 75
-
-Metrics/ParameterLists:
-  CountKeywordArgs: false
