--- workers/ruby/harness/Gemfile
@@ -1,12 +1,0 @@
-source 'https://rubygems.org'
-
-gemspec
-
-group :development do
-  gem 'grpc-tools', '~> 1.80'
-  gem 'minitest'
-  gem 'rake'
-  gem 'rbs', '~> 3.10'
-  gem 'rubocop', '~> 1.0'
-  gem 'steep', '~> 1.10'
-end
