+++ workers/ruby/harness/api/api_services_pb.rb
