+++ workers/ruby/harness/client.rb
