+++ workers/ruby/harness/helpers.rb
