+++ workers/ruby/harness/main.rb
