+++ workers/ruby/harness/worker.rb
