+++ workers/ruby/omes.gemspec
@@ -5,10 +5,11 @@ Gem::Specification.new do |s|
   s.authors = ['Temporal Technologies Inc']
   s.email = ['sdk@temporal.io']
   s.license = 'MIT'
-  s.files = Dir['*.rb'] + Dir['protos/**/*.rb']
+  s.files = Dir['*.rb'] + Dir['apps/**/*.rb'] + Dir['harness/**/*.rb'] +
+            Dir['protos/**/*.rb'] + Dir['sig/**/*.rbs'] + Dir['workerlib/**/*.rb']
   s.require_paths = ['.']
   s.add_dependency 'google-protobuf', '~> 4.0'
-  s.add_dependency 'harness', '~> 0.1'
+  s.add_dependency 'grpc', '~> 1.80'
   s.add_dependency 'temporalio', '~> 1.3'
   s.metadata['rubygems_mfa_required'] = 'true'
 end
