+++ workers/ruby/runner.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require 'bundler/setup'
-require 'harness'
-require_relative 'kitchen_sink_app'
+require_relative 'apps/registry'
 
-Harness.run(KitchenSinkApp.app) if __FILE__ == $PROGRAM_NAME
+Apps::Registry.run if __FILE__ == $PROGRAM_NAME
