+++ workers/ruby/sig/harness.rbs
