+++ workers/ruby/tests/test_worker.rb
