+++ workers/ruby/workerlib/kitchensink/activities.rb
